감상 목록과 감상 결과 UI 수정 반영
