+++ .idea/misc.xml
@@ -16,6 +16,8 @@
         <entry key="..\:/1_Allonsy_Project/.Git/Allons-y-front/Allons-y-front/app/src/main/res/layout/movie_collection_layout.xml" value="0.25364583333333335" />
         <entry key="..\:/1_Allonsy_Project/.Git/Allons-y-front/Allons-y-front/app/src/main/res/layout/movie_layout.xml" value="0.25364583333333335" />
         <entry key="..\:/1_Allonsy_Project/.Git/Allons-y-front/app/src/main/res/layout/activity_signup.xml" value="0.25" />
+        <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/drawable-v24/background_ellipse1.xml" value="0.1" />
+        <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/drawable-v24/background_ellipse2.xml" value="0.1" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/drawable/round_border_button.xml" value="0.1945" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/drawable/round_border_green.xml" value="0.1945" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/drawable/spinner_genre_background.xml" value="0.1945" />
@@ -30,12 +32,24 @@
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/dialog_spinkit.xml" value="0.16847826086956522" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/fragment_addreview.xml" value="0.17" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/fragment_result.xml" value="0.15" />
+        <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/fragment_result_ticket_back.xml" value="0.1" />
+        <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/fragment_result_ticket_front.xml" value="0.1" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/fragment_user_movie_list.xml" value="0.16847826086956522" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/movie_collection_layout.xml" value="0.16847826086956522" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/movie_layout.xml" value="0.16847826086956522" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/recyclerview_log.xml" value="0.16847826086956522" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/layout/recyclerview_row.xml" value="0.16847826086956522" />
         <entry key="..\:/Allons-y/HProject/Allons-y-front/app/src/main/res/menu/drawer_menu.xml" value="0.20364583333333333" />
+        <entry key="app/src/main/res/drawable-v24/background_ellipse1.xml" value="0.141" />
+        <entry key="app/src/main/res/drawable-v24/background_ellipse2.xml" value="0.141" />
+        <entry key="app/src/main/res/layout/activity_login.xml" value="0.223731884057971" />
+        <entry key="app/src/main/res/layout/activity_main2.xml" value="0.10739614994934144" />
+        <entry key="app/src/main/res/layout/activity_my_movie_list.xml" value="0.18931159420289856" />
+        <entry key="app/src/main/res/layout/fragment_result.xml" value="0.18931159420289856" />
+        <entry key="app/src/main/res/layout/fragment_result_ticket_back.xml" value="0.271286231884058" />
+        <entry key="app/src/main/res/layout/fragment_user_movie_list.xml" value="0.10739614994934144" />
+        <entry key="app/src/main/res/layout/movie_collection_layout.xml" value="0.23333333333333334" />
+        <entry key="app/src/main/res/layout/movie_layout.xml" value="0.23333333333333334" />
       </map>
     </option>
   </component>
