+++ app/build.gradle
@@ -45,6 +45,9 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     implementation("androidx.fragment:fragment-ktx:1.3.6")
 
+    // 이미지 평균색 추출
+    implementation 'androidx.palette:palette:1.0.0'
+
     // 로딩 다이얼로그
     implementation 'com.github.ybq:Android-SpinKit:1.4.0'
