+++ app/src/main/res/drawable-v24/background_ellipse2.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="oval">
+    <gradient
+        android:type="radial"
+        android:gradientRadius="50%p"
+        android:startColor="#66FE53BB"
+        android:endColor="#00000000" />
+    <size android:width="350dp" android:height="350dp"/>
+</shape>
